Imported NavbarModule to the demo app module
This enabled consumption of the library in the demo app for testing
and development purposes. The library was built into the dist directory
per Angular guidelines. Then an npm-link symbolic link was established,
obviating the need for publishing to the npm registry prematurely. The demo
app successfully displayed the mock sfl-navbar component.