First attempt at component customization module
This enables users of this library to customize component color beyond
the pre-built themes. Primary, accent, or warn presets from the theme can
be applied, or completely custom colors for the content (background coloring
not implemented yet). Customization is imported through customization
module which defines the custom color directive, ngsColor.